updated codes to a more sensible form per request

+++ Bangazon/Controllers/ProductsController.cs
@@ -163,42 +163,36 @@ namespace Bangazon.Controllers
         public async Task<IActionResult> RemoveCartProduct(int id)
         {
             
-            // Get a product from the DB that matches the id being passed into the method.
-            Product product = await _context.Product
-                                            .Where(p => p.ProductId == id)
-                                            .SingleOrDefaultAsync();
+            // Get a OrderProduct from the DB that matches the id being passed into the method.
+            OrderProduct orderProduct = await _context.OrderProduct
+                                            .Include(op => op.Product)
+                                            .Include(op => op.Order)
+                                            .SingleOrDefaultAsync(op => op.OrderProductId == id);
             
-            // Get the User for getting the current Order
+            // Get the User for getting the current Order to verify that it is the users order only
             ApplicationUser user = await GetCurrentUserAsync();
-            
 
-            // Get the order that is currently open, i.e. it does not have a date completed assigned to it or a paymentId
-            Order order = await _context.Order
-                                        .Include(o => o.OrderProducts)
-                                        .Where(o => o.UserId == user.Id && o.PaymentTypeId == null && o.DateCompleted == null)
-                                        .SingleOrDefaultAsync();
+            if (orderProduct.Order.UserId != user.Id) 
+            { 
+                return RedirectToAction("Index", "Home", new { id = orderProduct.Order.OrderId });
+            }
             
-            // Get the OrderProduct from the open order that matches the product id from above. 
-            // Take one incase there are multiples of the same product in the order
-            OrderProduct orderProduct = order.OrderProducts
-                                             .Where(op => op.OrderId == order.OrderId && op.ProductId == product.ProductId)
-                                                      .Take(1).SingleOrDefault();
 
             // Update the quantity of the product to increase by one since it is going back out of the cart.
-            product.Quantity = product.Quantity + 1;
-            _context.Product.Update(product);
+            orderProduct.Product.Quantity = orderProduct.Product.Quantity + 1;
+            _context.Product.Update(orderProduct.Product);
 
             // Remove the Order Product from the DB & remove it from the order from above
             _context.OrderProduct.Remove(orderProduct);
-            order.OrderProducts.Remove(orderProduct);
+            orderProduct.Order.OrderProducts.Remove(orderProduct);
 
             // created a bool to determine if the order is now product-less
             bool cancelOrder = false;
 
             // If order is product-less, remove the order from the DB and set bool to true
-            if (order.OrderProducts.Count == 0) 
+            if (orderProduct.Order.OrderProducts.Count == 0) 
             {
-                _context.Order.Remove(order);
+                _context.Order.Remove(orderProduct.Order);
                 cancelOrder = true;
             }
 
@@ -212,7 +206,7 @@ namespace Bangazon.Controllers
             }
             else 
             { 
-                return RedirectToAction("Details", "Orders", new { id = order.OrderId });
+                return RedirectToAction("Details", "Orders", new { id = orderProduct.Order.OrderId });
             }
 
         }
